removed wrong files added right

+++ zadania1/zad4/main.c
@@ -3,11 +3,11 @@
 #include <math.h>
 
 int main(){
-    int a,b,c;
+    double a,b,c;
     char znak;
     do{
         printf("Podaj 3 liczby: ");
-        scanf("%d %d %d",&a,&b,&c);
+        scanf("%lf %lf %lf",&a,&b,&c);
         if(a>0 && b>0 && c>0){
             if(a>=b && a>=c){
                 if((b+c)>a) printf("To moze byc trojkat");
